Major change to locale handling.  Previously we set LC_NUMERIC=C because libnyquist failed otherwise.  However, this caused two problems:
 - Audacity did not use the user's preferred numeric locale when
   displaying numbers on the screen.

 - When other code (such as LADSPA plugins) called setlocale(),
   it broke libnyquist again.

This patch causes Audacity to keep the user's default locale.  Numbers
are always displayed in the preferred local format, except when printing
to XML files.

To work around the libnyquist problem, we temporarily set LC_NUMERIC=C
only when calling libnyquist.  (This is not thread-safe.  Ideally,
nyquist should be fixed to work in any locale.)


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@3646 47890b92-0858-11df-a26f-8b716316a5bc