Offset fixes to TrackPanel and WaveTrack;
revert Wavetrack::Set and WaveTrack::Get to original behavior with
respect to offsets.  It was more consistent and Get was used in other places that I missed.

Correct TrackPanel to use WaveTrack with absolute times (as above) and
not try to apply offsets.  As it turns out, this also corrected
several problem calls into the Envelope code which have the same
offset behavior as the reverted WaveTrack code.

This also corrects the offset playback bug I introduced in my last
patch.

At fully zoomed in levels, envelopes are not displaying entirely
consistently with set values; the control points and display are not
completely matching the effect on points.  This appears to be
unrelated to the above bug, but I'm still tracking it down.  This bug
is not due to absolute offset, but a finer sub-sample-width alignment
error.


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@3954 47890b92-0858-11df-a26f-8b716316a5bc