The patch thus introduces keyboard shortcuts for making short/long jumps while paused and while playing. However,it doesnt affect any of the current functionality (using (shift+)arrow to jump while playing). Keyboard prefs are bound to , and . for short jumps, and with shift (i.e. <,>) for long jumps.
Also added new function to adjust boundaries, so that keyboard
shortcuts for this can be assigned more flexibly. E.g. the new preference
allows the same key to be used for setting the left boundary
while playing, and extending that boundary while stopped.

Changes contained in two functions in TrackPanel.cpp, and some keyboard
preferences in Menus.cpp. This present code just uses one function for both
left and right movements, and is less tied to keyboard shortcuts than the older
code. At the moment, it just duplicated functionality of older code, but it could
replace older code in future releases.


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@6782 47890b92-0858-11df-a26f-8b716316a5bc