Maintain track position (relative to start of tracklist) information and post resize/update events to inform "users" of changes.
Created a VisibleTrackIterator and updated TrackPanel, TrackArtist, and AudacityProject to take advantage of it.

Do not read preference values during redraw of spectrum display types since it can be slow.  They are now read when preferences are updated and written during termination.  This required changes to TrackPanel as well since they two classes were basically using the config file as a means to pass min/max values.

Moved the vruler updating code from DrawVRuler to UpdateVRuler to support the dynamic ruler sizing.  This was necessary to correctly determine the needed width before actually starting to draw the vruler.  On the Mac, the ruler would often draw incorrectly and on Windows you could see it draw at one size and then expand to another.

Sliding clips no longer slides it to the wrong track.

Added InvalidateSpectrumCache() to TrackArtist to provide consistent spectrum cache rebuilding.  This helps to fix a bug where the cache of 2nd stereo track wasn't being invalidated.

A few cases where a full refresh of the TrackPanel wasn't needed were change to only refresh the modified track(s).

The track title is no longer "pushed" for all tracks when the track menu is displayed.

The track minimize and close buttons are no longer "pushed" for all tracks when moving scrolling while one is pushed.

The left offset for the AdornedRuler was off by one causing the region, playback, and record indicators to not line up with the track indicator.

The vertical ruler is now right-aligned to prevent text overflowing into the track label area.

When switching between track display types, the highlighted region in the horizontal ruler would not be aligned with the highlighted region in the tracks.  This is now corrected.

When cursoring down, scrolling no longer jumps the cursor back to the top of the window so you no longer have to keep looking from the bottom to the top.

A few 1-off drawing issues were also resolved.


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@8963 47890b92-0858-11df-a26f-8b716316a5bc