Converted all wxDC::DrawLine()s to AColor::Line()s so that it can handle the line drawing differences among platforms.
Hopefully fixed all 1-off drawing issues that creaped in due to the platform
differences.  Everyone should keep an extra critical eye open for stray
pixels or lines that seem to be longer at one of the ends.

Reworked much of the waveform drawing code in TrackArtist to improve
performance.

Due to the performance gains in TrackArtist, the Mac specific waveform
drawing code was removed since it was out of sync with the other platforms.

Fixed drawing and click detection within NoteTrack label area.


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@9112 47890b92-0858-11df-a26f-8b716316a5bc