Adding label synchronisation for Change Speed, Change Tempo, and Time Scale (SBSMS).
This will move the labels in any *selected* label track. This currently does
not respect track linking/groups, regardless of whether it is enabled. There is
also a certain amount of duplicated code because these effects could do with
some refactoring, but it's not worth disturbing things at the moment.

It might or might not be possible to make it behave slightly oddly in some
cases, thanks to the general linking state of confusion, but this should be
better than before as it's now at least possible to use these effects without
effectively losing the label timing information.


git-svn-id: http://audacity.googlecode.com/svn/sf-cvs/trunk/audacity-src@9997 47890b92-0858-11df-a26f-8b716316a5bc