(bug 20, P2)
Changed the bSilentlyCorrect parameter to DirManager::ProjectFSCK() to be bAutoRecoverMode. If project check errors are found during auto-recovery, it's hazardous to allow the user the various options available during normal project open. But rather than squelching the messages as well as option dialogs during auto-recover, show the messages in the log and bring up an alert. 

This means the helper methods Find*() no longer need bSilentlyCorrect -- they always report to the log. 

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@10677 47890b92-0858-11df-a26f-8b716316a5bc