(Fix Bug 250 - Pan sliders sometimes not drawn in Mixer Board, P2)
This bug was introduced in commit 10680 and caused in MixerTrackCluster::UpdatePan(), by using USE_MIDI (which is on in releases) instead of EXPERIMENTAL_MIDI_OUT (which is off). It checked whether a NoteTrack pointer was NULL, and if not, hid the Pan slider (not used in NoteTrack clusters). But an invalid pointer is non-null, so was causing the Pan slider to always be hidden.

git-svn-id: http://audacity.googlecode.com/svn/audacity-src/trunk@10772 47890b92-0858-11df-a26f-8b716316a5bc